requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
akracer==0.0.13
akshare==1.16.81
amqp==5.3.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.2.1
et_xmlfile==2.0.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
greenlet==3.2.1
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpath==0.82.2
kombu==5.5.3
lxml==5.4.0
MarkupSafe==3.0.2
multidict==6.4.3
nest-asyncio==1.6.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
prompt_toolkit==3.0.51
propcache==0.3.1
py-mini-racer==0.6.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.31.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.40
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
xlrd==2.0.1
yarl==1.20.0
gunicorn
cryptography